renaming template package
