+++ examples/web/main.go
@@ -31,7 +31,7 @@ import (
 	"github.com/ardanlabs/kronk/sdk/kronk"
 	"github.com/ardanlabs/kronk/sdk/kronk/defaults"
 	"github.com/ardanlabs/kronk/sdk/kronk/model"
-	"github.com/ardanlabs/kronk/sdk/kronk/templater"
+	"github.com/ardanlabs/kronk/sdk/kronk/template"
 	"github.com/ardanlabs/kronk/sdk/tools/catalog"
 	"github.com/ardanlabs/kronk/sdk/tools/libs"
 	"github.com/ardanlabs/kronk/sdk/tools/models"
@@ -102,10 +102,9 @@ func run() error {
 		return fmt.Errorf("unable to init kronk: %w", err)
 	}
 
-	krnChat, err := kronk.New(modelInstances, model.Config{
+	krnChat, err := kronk.New(modelInstances, template.New(), model.Config{
 		Log:       kronk.FmtLogger,
 		ModelFile: info.ModelFile,
-		Templater: templater.New(),
 		NBatch:    32 * 1024,
 	})
 	if err != nil {
