+++ sdk/kronk/tests/media_test.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/ardanlabs/kronk/sdk/kronk"
 	"github.com/ardanlabs/kronk/sdk/kronk/model"
-	"github.com/ardanlabs/kronk/sdk/kronk/templater"
+	"github.com/ardanlabs/kronk/sdk/kronk/template"
 	"github.com/ardanlabs/kronk/sdk/tools/models"
 	"github.com/google/uuid"
 	"golang.org/x/sync/errgroup"
@@ -143,10 +143,9 @@ func initMediaTest(t *testing.T, mp models.Path, mediaFile string) (*kronk.Kronk
 
 	// -------------------------------------------------------------------------
 
-	krn, err := kronk.New(modelInstances, model.Config{
+	krn, err := kronk.New(modelInstances, template.New(), model.Config{
 		ModelFile: mp.ModelFile,
 		ProjFile:  mp.ProjFile,
-		Templater: templater.New(),
 	})
 
 	if err != nil {
